requirements.txt
torch>=2.0.0
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
transformers==4.48.0
hydra-core==1.3.2
pytorch-lightning==2.5.0